+++ apps/dashboard/src/app/knowledge-box/knowledge-box-training/knowledge-box-training.component.spec.ts
@@ -3,7 +3,7 @@ import { SDKService, STFTrackingService, TranslatePipeMock } from '@flaps/core';
 import { TrainingStatus } from '@nuclia/core';
 import { of } from 'rxjs';
 
-import { KnowledgeBoxProcessesComponent } from './knowledge-box-processes.component';
+import { KnowledgeBoxTrainingComponent } from './knowledge-box-training.component';
 import { MockModule } from 'ng-mocks';
 import {
   PaButtonModule,
@@ -15,11 +15,10 @@ import {
 } from '@guillotinaweb/pastanaga-angular';
 import { DropdownButtonComponent } from '@nuclia/sistema';
 import { SvgIconRegistryService } from 'angular-svg-icon';
-import { STFSectionNavbarModule } from '../../components/section-navbar';
 
-describe('KnowledgeBoxProcessComponent', () => {
-  let component: KnowledgeBoxProcessesComponent;
-  let fixture: ComponentFixture<KnowledgeBoxProcessesComponent>;
+describe('KnowledgeBoxTrainingComponent', () => {
+  let component: KnowledgeBoxTrainingComponent;
+  let fixture: ComponentFixture<KnowledgeBoxTrainingComponent>;
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
@@ -30,10 +29,9 @@ describe('KnowledgeBoxProcessComponent', () => {
         MockModule(PaTogglesModule),
         MockModule(PaIconModule),
         MockModule(PaPopupModule),
-        MockModule(STFSectionNavbarModule),
         DropdownButtonComponent,
       ],
-      declarations: [KnowledgeBoxProcessesComponent, TranslatePipeMock],
+      declarations: [KnowledgeBoxTrainingComponent, TranslatePipeMock],
       providers: [
         {
           provide: SDKService,
@@ -54,7 +52,7 @@ describe('KnowledgeBoxProcessComponent', () => {
   });
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(KnowledgeBoxProcessesComponent);
+    fixture = TestBed.createComponent(KnowledgeBoxTrainingComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });